Models refactored & cleanup ★
