+++ api/Middlewares/check-auth.js
@@ -1,11 +1,12 @@
 const jwt = require('jsonwebtoken');
 const responseHandler = require('../Utils/responseHandler');
 
-exports.checkStudentAuth = function (req, res, next) {
+// eslint-disable-next-line consistent-return
+exports.checkStudentAuth = (req, res, next) => {
   try {
     const token = req.headers.authorization.split(' ')[1];
     const decoded = jwt.verify(token, 'secret');
-    if (decoded.user_role == 'student') {
+    if (decoded.user_role === 'student') {
       req.userData = decoded;
       next();
     } else {
@@ -15,11 +16,11 @@ exports.checkStudentAuth = function (req, res, next) {
     return responseHandler.error(res, 'Unauthorized', 401);
   }
 };
-exports.checkProfessorAuth = function (req, res, next) {
+exports.checkProfessorAuth = (req, res, next) => {
   try {
     const token = req.headers.authorization.split(' ')[1];
     const decoded = jwt.verify(token, 'secret');
-    if (decoded.user_role == 'professor') {
+    if (decoded.user_role === 'professor') {
       req.userData = decoded;
       next();
     } else {
